requirements.txt
openai-whisper
ffmpeg-python
azure-cognitiveservices-speech
